+++ app/Http/Controllers/api/UserController.php
@@ -2,6 +2,7 @@
 
 namespace App\Http\Controllers\api;
 
+use App\Http\Controllers\helpers\RoleAuthorizationHelper;
 use Illuminate\Routing\Controller;
 use App\Http\Requests\UpdateUserRequest;
 use App\Http\Resources\UserResource;
@@ -12,21 +13,13 @@ use Illuminate\Support\Facades\Auth;
 
 class UserController extends Controller
 {
+    protected $roleAuthorizationHelper;
+
     public function __construct()
     {
         $this->middleware('auth:sanctum');
+        $this->roleAuthorizationHelper = new RoleAuthorizationHelper();
     }
-
-    /**
-     * Display a listing of the resource.
-     */
-    public function index()
-    {
-        $users = User::orderBy('id', 'desc');
-
-        return new UserCollection($users->get());
-    }
-
 
     public function enterpriseUsers(Request $request)
     {
@@ -36,42 +29,37 @@ class UserController extends Controller
 
         return new UserCollection($users->get());
     }
-
-    /**
-     * Store a newly created resource in storage.
-     */
-    public function store(UpdateUserRequest $request)
-    {
-        //
-    }
-
-    /**
-     * Display the specified resource.
-     */
-    public function show(string $id)
-    {
-        //
-    }
-
-    /**
-     * Update the specified resource in storage.
-     */
-    public function update(Request $request, string $id)
-    {
-        //
-    }
-
+  
     /**
      * Remove the specified resource from storage.
      */
     public function destroy(User $user)
     {
+        if (!$this->roleAuthorizationHelper->hasPermission($user->role, 'user.delete')) {
+            return response()->json([
+                'message' => 'No estás autorizado para realizar esta acción.'
+            ], 401);
+        }
+
         $user->delete();
         return response(null, 204);
     }
 
     public function showProfile()
     {
+        $user = User::find(Auth::user()->id);
+        if (!$user) {
+            return response()->json([
+                'message' => 'Usuario no encontrado.'
+            ], 404);
+        }
+
+        if (!$this->roleAuthorizationHelper->hasPermission($user->role, 'user.read')) {
+            return response()->json([
+                'message' => 'No estás autorizado para realizar esta acción.'
+            ], 401);
+        }
+
         return response()->json([
             'data' => UserResource::make(Auth::user())
         ], 200);
@@ -80,11 +68,16 @@ class UserController extends Controller
     public function updateProfile(UpdateUserRequest $request)
     {
         $user = User::find(Auth::user()->id);
-
         if (!$user) {
             return response()->json([
-                'message' => 'User not found'
+                'message' => 'Usuario no encontrado.'
             ], 404);
+        }
+
+        if (!$this->roleAuthorizationHelper->hasPermission($user->role, 'user.update')) {
+            return response()->json([
+                'message' => 'No estás autorizado para realizar esta acción.'
+            ], 401);
         }
 
         $user->update($request->validated());
@@ -99,8 +92,44 @@ class UserController extends Controller
         $user = Auth::user();
         $enterprise = $user->enterprise;
 
+        if (!$user) {
+            return response()->json([
+                'message' => 'Usuario no encontrado.'
+            ], 404);
+        }
+
+        if (!$this->roleAuthorizationHelper->hasPermission($user->role, 'enterprise.read')) {
+            return response()->json([
+                'message' => 'No estás autorizado para realizar esta acción.'
+            ], 401);
+        }
+
         return response()->json([
             'data' => $enterprise
         ], 200);
     }
+
+    // /**
+    //  * Store a newly created resource in storage.
+    //  */
+    // public function store(UpdateUserRequest $request)
+    // {
+    //     //
+    // }
+
+    // /**
+    //  * Display the specified resource.
+    //  */
+    // public function show(string $id)
+    // {
+    //     //
+    // }
+
+    // /**
+    //  * Update the specified resource in storage.
+    //  */
+    // public function update(Request $request, string $id)
+    // {
+    //     //
+    // }
 }
